[Base] Move implementation out of header where appropriate, and qualify std namespace.
